requirements.txt
requests>=2.31.0
pymongo>=4.6.1
python-cloudflare>=2.8.15
openai>=1.12.0
typing-extensions>=4.9.0
python-dotenv>=1.0.0
